v1.2.1
This is quite a large internal change which doesn't affect the API, but does include various bugfixes.

v1.2.1 is not v1.0.4 (which would be logical following the previous git tag of v1.0.3) due to an inconsistency of unknown origin in our version numbers. We have decided to follow the previous larger of the two version numbers which was in src/mux-analytics.brs. Moving forward the two will be kept in sync to avoid confusion.

The build and test process, historically based on gulp, has been replaced with a custom golang tool "test_runner" which does superficially similar things. The tasks are defined within test_runner and remain much the same as before. To run test_runner you must build it with golang by running `go build` in the project directory, then run the resulting executable. More details are in README.md. This change was made as keeping the gulp setup up to date in a semi secure way was becoming too problematic.

Linting no longer uses the BSkyB linter but instead that from rokucommunity.

The tests are back up and running if you have a real device configured via local.properties. (test_runner will load this file and adjust the target device accordingly).

The tests uncovered some miscellaneous bugs which are fixed so all tests pass.

New tests were written to account for a reported problem with Seeking events. These were also fixed so that the tests pass.